backend/requirements.txt
Django==4.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
pyshp==2.3.1